폴더 및 파일 구조 수정 - actions/profile.tsx
